Unify DPDA and NPDA.
